+++ pyinjective/core/network.py
@@ -2,7 +2,10 @@ import datetime
 from abc import ABC, abstractmethod
 from http.cookies import SimpleCookie
 from typing import List, Optional
+from warnings import warn
 
+import grpc
+from grpc import ChannelCredentials
 from grpc.aio import Call, Metadata
 
 
@@ -115,8 +118,20 @@ class Network:
         chain_cookie_assistant: CookieAssistant,
         exchange_cookie_assistant: CookieAssistant,
         explorer_cookie_assistant: CookieAssistant,
-        use_secure_connection: bool = False,
+        use_secure_connection: Optional[bool] = None,
+        grpc_channel_credentials: Optional[ChannelCredentials] = None,
+        grpc_exchange_channel_credentials: Optional[ChannelCredentials] = None,
+        grpc_explorer_channel_credentials: Optional[ChannelCredentials] = None,
+        chain_stream_channel_credentials: Optional[ChannelCredentials] = None,
     ):
+        # the `use_secure_connection` parameter is ignored and will be deprecated soon.
+        if use_secure_connection is not None:
+            warn(
+                "use_secure_connection parameter in Network is no longer used and will be deprecated",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+
         self.lcd_endpoint = lcd_endpoint
         self.tm_websocket_endpoint = tm_websocket_endpoint
         self.grpc_endpoint = grpc_endpoint
@@ -129,7 +144,10 @@ class Network:
         self.chain_cookie_assistant = chain_cookie_assistant
         self.exchange_cookie_assistant = exchange_cookie_assistant
         self.explorer_cookie_assistant = explorer_cookie_assistant
-        self.use_secure_connection = use_secure_connection
+        self.grpc_channel_credentials = grpc_channel_credentials
+        self.grpc_exchange_channel_credentials = grpc_exchange_channel_credentials
+        self.grpc_explorer_channel_credentials = grpc_explorer_channel_credentials
+        self.chain_stream_channel_credentials = chain_stream_channel_credentials
 
     @classmethod
     def devnet(cls):
@@ -157,6 +175,11 @@ class Network:
         if node not in nodes:
             raise ValueError("Must be one of {}".format(nodes))
 
+        grpc_channel_credentials = grpc.ssl_channel_credentials()
+        grpc_exchange_channel_credentials = grpc.ssl_channel_credentials()
+        grpc_explorer_channel_credentials = grpc.ssl_channel_credentials()
+        chain_stream_channel_credentials = grpc.ssl_channel_credentials()
+
         if node == "lb":
             lcd_endpoint = "https://testnet.sentry.lcd.injective.network:443"
             tm_websocket_endpoint = "wss://testnet.sentry.tm.injective.network:443/websocket"
@@ -167,7 +190,6 @@ class Network:
             chain_cookie_assistant = BareMetalLoadBalancedCookieAssistant()
             exchange_cookie_assistant = BareMetalLoadBalancedCookieAssistant()
             explorer_cookie_assistant = BareMetalLoadBalancedCookieAssistant()
-            use_secure_connection = True
         else:
             lcd_endpoint = "https://testnet.lcd.injective.network:443"
             tm_websocket_endpoint = "wss://testnet.tm.injective.network:443/websocket"
@@ -178,7 +200,6 @@ class Network:
             chain_cookie_assistant = DisabledCookieAssistant()
             exchange_cookie_assistant = DisabledCookieAssistant()
             explorer_cookie_assistant = DisabledCookieAssistant()
-            use_secure_connection = True
 
         return cls(
             lcd_endpoint=lcd_endpoint,
@@ -193,7 +214,10 @@ class Network:
             chain_cookie_assistant=chain_cookie_assistant,
             exchange_cookie_assistant=exchange_cookie_assistant,
             explorer_cookie_assistant=explorer_cookie_assistant,
-            use_secure_connection=use_secure_connection,
+            grpc_channel_credentials=grpc_channel_credentials,
+            grpc_exchange_channel_credentials=grpc_exchange_channel_credentials,
+            grpc_explorer_channel_credentials=grpc_explorer_channel_credentials,
+            chain_stream_channel_credentials=chain_stream_channel_credentials,
         )
 
     @classmethod
@@ -213,7 +237,10 @@ class Network:
         chain_cookie_assistant = BareMetalLoadBalancedCookieAssistant()
         exchange_cookie_assistant = BareMetalLoadBalancedCookieAssistant()
         explorer_cookie_assistant = BareMetalLoadBalancedCookieAssistant()
-        use_secure_connection = True
+        grpc_channel_credentials = grpc.ssl_channel_credentials()
+        grpc_exchange_channel_credentials = grpc.ssl_channel_credentials()
+        grpc_explorer_channel_credentials = grpc.ssl_channel_credentials()
+        chain_stream_channel_credentials = grpc.ssl_channel_credentials()
 
         return cls(
             lcd_endpoint=lcd_endpoint,
@@ -228,7 +255,10 @@ class Network:
             chain_cookie_assistant=chain_cookie_assistant,
             exchange_cookie_assistant=exchange_cookie_assistant,
             explorer_cookie_assistant=explorer_cookie_assistant,
-            use_secure_connection=use_secure_connection,
+            grpc_channel_credentials=grpc_channel_credentials,
+            grpc_exchange_channel_credentials=grpc_exchange_channel_credentials,
+            grpc_explorer_channel_credentials=grpc_explorer_channel_credentials,
+            chain_stream_channel_credentials=chain_stream_channel_credentials,
         )
 
     @classmethod
@@ -246,7 +276,6 @@ class Network:
             chain_cookie_assistant=DisabledCookieAssistant(),
             exchange_cookie_assistant=DisabledCookieAssistant(),
             explorer_cookie_assistant=DisabledCookieAssistant(),
-            use_secure_connection=False,
         )
 
     @classmethod
@@ -263,8 +292,20 @@ class Network:
         chain_cookie_assistant: Optional[CookieAssistant] = None,
         exchange_cookie_assistant: Optional[CookieAssistant] = None,
         explorer_cookie_assistant: Optional[CookieAssistant] = None,
-        use_secure_connection: bool = False,
+        use_secure_connection: Optional[bool] = None,
+        grpc_channel_credentials: Optional[ChannelCredentials] = None,
+        grpc_exchange_channel_credentials: Optional[ChannelCredentials] = None,
+        grpc_explorer_channel_credentials: Optional[ChannelCredentials] = None,
+        chain_stream_channel_credentials: Optional[ChannelCredentials] = None,
     ):
+        # the `use_secure_connection` parameter is ignored and will be deprecated soon.
+        if use_secure_connection is not None:
+            warn(
+                "use_secure_connection parameter in Network is no longer used and will be deprecated",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+
         chain_assistant = chain_cookie_assistant or DisabledCookieAssistant()
         exchange_assistant = exchange_cookie_assistant or DisabledCookieAssistant()
         explorer_assistant = explorer_cookie_assistant or DisabledCookieAssistant()
@@ -281,8 +322,59 @@ class Network:
             chain_cookie_assistant=chain_assistant,
             exchange_cookie_assistant=exchange_assistant,
             explorer_cookie_assistant=explorer_assistant,
-            use_secure_connection=use_secure_connection,
+            grpc_channel_credentials=grpc_channel_credentials,
+            grpc_exchange_channel_credentials=grpc_exchange_channel_credentials,
+            grpc_explorer_channel_credentials=grpc_explorer_channel_credentials,
+            chain_stream_channel_credentials=chain_stream_channel_credentials,
+        )
+
+    @classmethod
+    def custom_chain_and_public_indexer_mainnet(
+        cls,
+        lcd_endpoint,
+        tm_websocket_endpoint,
+        grpc_endpoint,
+        chain_stream_endpoint,
+        chain_cookie_assistant: Optional[CookieAssistant] = None,
+    ):
+        mainnet_network = cls.mainnet()
+
+        return cls.custom(
+            lcd_endpoint=lcd_endpoint,
+            tm_websocket_endpoint=tm_websocket_endpoint,
+            grpc_endpoint=grpc_endpoint,
+            grpc_exchange_endpoint=mainnet_network.grpc_exchange_endpoint,
+            grpc_explorer_endpoint=mainnet_network.grpc_explorer_endpoint,
+            chain_stream_endpoint=chain_stream_endpoint,
+            chain_id="injective-1",
+            env="mainnet",
+            chain_cookie_assistant=chain_cookie_assistant or DisabledCookieAssistant(),
+            exchange_cookie_assistant=mainnet_network.exchange_cookie_assistant,
+            explorer_cookie_assistant=mainnet_network.explorer_cookie_assistant,
+            grpc_channel_credentials=None,
+            grpc_exchange_channel_credentials=mainnet_network.grpc_exchange_channel_credentials,
+            grpc_explorer_channel_credentials=mainnet_network.grpc_explorer_channel_credentials,
+            chain_stream_channel_credentials=None,
         )
 
     def string(self):
         return self.env
+
+    def create_chain_grpc_channel(self) -> grpc.Channel:
+        return self._create_grpc_channel(self.grpc_endpoint, self.grpc_channel_credentials)
+
+    def create_exchange_grpc_channel(self) -> grpc.Channel:
+        return self._create_grpc_channel(self.grpc_exchange_endpoint, self.grpc_exchange_channel_credentials)
+
+    def create_explorer_grpc_channel(self) -> grpc.Channel:
+        return self._create_grpc_channel(self.grpc_explorer_endpoint, self.grpc_explorer_channel_credentials)
+
+    def create_chain_stream_grpc_channel(self) -> grpc.Channel:
+        return self._create_grpc_channel(self.chain_stream_endpoint, self.chain_stream_channel_credentials)
+
+    def _create_grpc_channel(self, endpoint: str, credentials: Optional[ChannelCredentials]) -> grpc.Channel:
+        if credentials is None:
+            channel = grpc.aio.insecure_channel(endpoint)
+        else:
+            channel = grpc.aio.secure_channel(endpoint, credentials)
+        return channel