+++ tests/test_async_client_deprecation_warnings.py
@@ -1,5 +1,6 @@
 from warnings import catch_warnings
 
+import grpc
 import pytest
 
 from pyinjective.async_client import AsyncClient
@@ -586,7 +587,7 @@ class TestAsyncClientDeprecationWarnings:
         assert str(deprecation_warnings[0].message) == "This method is deprecated. Use fetch_oracle_price instead"
 
     @pytest.mark.asyncio
-    async def test_stream_keepalive_deprecation_warning(
+    async def test_stream_oracle_prices_deprecation_warning(
         self,
         oracle_servicer,
     ):
@@ -1707,3 +1708,17 @@ class TestAsyncClientDeprecationWarnings:
         deprecation_warnings = [warning for warning in all_warnings if issubclass(warning.category, DeprecationWarning)]
         assert len(deprecation_warnings) == 1
         assert str(deprecation_warnings[0].message) == "This method is deprecated. Use fetch_latest_block instead"
+
+    def test_credentials_parameter_deprecation_warning(
+        self,
+        auth_servicer,
+    ):
+        with catch_warnings(record=True) as all_warnings:
+            AsyncClient(network=Network.local(), credentials=grpc.ssl_channel_credentials())
+
+        deprecation_warnings = [warning for warning in all_warnings if issubclass(warning.category, DeprecationWarning)]
+        assert len(deprecation_warnings) == 1
+        assert (
+            str(deprecation_warnings[0].message)
+            == "credentials parameter in AsyncClient is no longer used and will be deprecated"
+        )